Fix for #21 - break each dataset into its own directory, control each package's imported structure.
