+++ nidata/anatomical/oasis_vbm/__init__.py
@@ -1,0 +1,1 @@
+from .datasets import *
