+++ nidata/atlas/__init__.py
@@ -1,2 +1,3 @@
-from .datasets import *
-from .haxby_etal_2011 import *
+import os.path as _osp
+from ..core._utils import import_all_submodules as _impall
+_impall(_osp.dirname(_osp.abspath(__file__)), locals(), globals())
