+++ nidata/atlas/craddock_2012/craddock_2012.rst
