+++ nidata/atlas/haxby_etal_2011/__init__.py
@@ -1,0 +1,1 @@
+from .datasets import *
