+++ nidata/atlas/haxby_etal_2011/datasets.py
@@ -1,5 +1,7 @@
-from ..core.fetchers import HttpFetcher
-from ..core.datasets import Dataset
+"""
+"""
+from ...core.fetchers import HttpFetcher
+from ...core.datasets import Dataset
 
 
 class HaxbyEtal2011(Dataset):
