+++ nidata/atlas/icbm152_2009/__init__.py
@@ -1,0 +1,1 @@
+from .datasets import *
