+++ nidata/atlas/mni152_template/__init__.py
@@ -1,0 +1,1 @@
+from .datasets import *
