+++ nidata/atlas/mni152_template/datasets.py
@@ -1,0 +1,61 @@
+# *- encoding: utf-8 -*-
+"""
+Utilities to download NeuroImaging-based atlases
+"""
+# Author: Alexandre Abraham, Philippe Gervais
+# License: simplified BSD
+
+import contextlib
+import collections
+import os
+import tarfile
+import zipfile
+import sys
+import shutil
+import time
+import hashlib
+import fnmatch
+import warnings
+import re
+import base64
+
+import numpy as np
+from scipy import ndimage
+from sklearn.datasets.base import Bunch
+
+from ...core._utils.compat import (_basestring, BytesIO, cPickle, _urllib,
+                                   md5_hash)
+from ...core._utils.niimg import check_niimg, new_img_like
+from ...core.fetchers import (format_time, md5_sum_file, fetch_files,
+                              get_dataset_dir, get_dataset_descr)
+
+
+def fetch_mni152_template():
+    """Load skullstripped 2mm version of the MNI152 originally distributed
+    with FSL
+
+    Returns
+    -------
+    mni152_template: nibabel object corresponding to the template
+
+
+    References
+    ----------
+
+    VS Fonov, AC Evans, K Botteron, CR Almli, RC McKinstry, DL Collins and
+    BDCG, Unbiased average age-appropriate atlases for pediatric studies,
+    NeuroImage, Volume 54, Issue 1, January 2011, ISSN 1053–8119, DOI:
+    10.1016/j.neuroimage.2010.07.033
+
+    VS Fonov, AC Evans, RC McKinstry, CR Almli and DL Collins, Unbiased
+    nonlinear average age-appropriate brain templates from birth to adulthood,
+    NeuroImage, Volume 47, Supplement 1, July 2009, Page S102 Organization for
+    Human Brain Mapping 2009 Annual Meeting, DOI: 10.1016/S1053-8119(09)70884-5
+
+    """
+    # https://raw.githubusercontent.com/nilearn/nilearn/master/nilearn/data/avg152T1_brain.nii.gz
+    package_directory = os.path.dirname(os.path.abspath(__file__))
+    path = os.path.join(package_directory, "data", "avg152T1_brain.nii.gz")
+
+    # XXX Should we load the image here?
+    return check_niimg(path)
