+++ nidata/atlas/smith_2009/smith_2009.rst
