+++ nidata/atlas/yeo/__init__.py
@@ -1,0 +1,1 @@
+from .datasets import *
