+++ nidata/core/_utils/__init__.py
@@ -1,0 +1,1 @@
+from .importing import import_all_submodules
