+++ nidata/core/_utils/importing.py
@@ -1,0 +1,17 @@
+"""
+"""
+import os.path
+
+
+def import_all_submodules(dir_path, locals, globals, recursive=True):
+
+    subdirs = filter(lambda fil: os.path.isdir(os.path.join(dir_path, fil)),
+                     os.listdir(dir_path))
+
+    for subdir in subdirs:
+        print subdir
+        if recursive:
+            exec('from .%s import *' % subdir) in locals, globals
+        else:
+            exec('from . import %s' % subdir) in locals, globals
+
