+++ nidata/functional/haxby_etal_2001/__init__.py
@@ -1,0 +1,1 @@
+from .datasets import *
