+++ nidata/functional/miyawaki_2008/miyawaki_2008.rst
