+++ nidata/localizer/brainomics/datasets.py
@@ -23,9 +23,9 @@ import numpy as np
 from scipy import ndimage
 from sklearn.datasets.base import Bunch
 
-from ..core._utils.compat import _basestring, BytesIO, cPickle, _urllib, md5_hash
-from ..core._utils.niimg import check_niimg, new_img_like
-from ..core.fetchers import format_time, md5_sum_file, fetch_files, get_dataset_dir, get_dataset_descr
+from ...core._utils.compat import _basestring, BytesIO, cPickle, _urllib, md5_hash
+from ...core._utils.niimg import check_niimg, new_img_like
+from ...core.fetchers import format_time, md5_sum_file, fetch_files, get_dataset_dir, get_dataset_descr
 
 
 def fetch_localizer_contrasts(contrasts, n_subjects=None, get_tmaps=False,
