+++ nidata/resting_state/ABIDE_pcp/__init__.py
@@ -1,0 +1,1 @@
+from .datasets import fetch_abide_pcp
