+++ nidata/resting_state/adhd/__init__.py
@@ -1,0 +1,1 @@
+from .datasets import fetch_adhd
