+++ nidata/resting_state/nyu/__init__.py
@@ -1,0 +1,1 @@
+from .datasets import fetch_nyu_rest
