+++ nidata/resting_state/nyu/nyu.rst
